backend/pyproject.toml
[project]
name = "pt-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "Phong Le",email = "phong.le@reactron.dev"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "strawberry-graphql (>=0.258.0,<0.259.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "websockets (>=15.0,<16.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[tool.ruff]
line-length = 88  # Match Black’s default
target-version = "py311"  # Set your Python version

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B"]  # Enable rules (E: pycodestyle, F: Pyflakes, I: isort, etc.)
ignore = ["B008"]  # Ignore specific rules if needed

[tool.ruff.format]
quote-style = "double"  # Use double quotes for strings
indent-style = "space"  # Use spaces instead of tabs
line-ending = "lf"  # Ensure line endings are LF (Unix)
